--- a/src/tigerbeetle/cli.zig
+++ b/src/tigerbeetle/cli.zig
@@ -45,8 +45,8 @@ const CliArgs = union(enum) {
         limit_storage: flags.ByteSize = .{ .bytes = constants.storage_size_limit_max },
         cache_accounts: flags.ByteSize = .{ .bytes = constants.cache_accounts_size_default },
         cache_transfers: flags.ByteSize = .{ .bytes = constants.cache_transfers_size_default },
-        cache_transfers_posted: flags.ByteSize =
-            .{ .bytes = constants.cache_transfers_posted_size_default },
+        cache_transfers_pending: flags.ByteSize =
+            .{ .bytes = constants.cache_transfers_pending_size_default },
         cache_account_history: flags.ByteSize =
             .{ .bytes = constants.cache_account_history_size_default },
         cache_grid: flags.ByteSize = .{ .bytes = constants.grid_cache_size_default },
@@ -73,7 +73,7 @@ const CliArgs = union(enum) {
     benchmark: struct {
         cache_accounts: ?[]const u8 = null,
         cache_transfers: ?[]const u8 = null,
-        cache_transfers_posted: ?[]const u8 = null,
+        cache_transfers_pending: ?[]const u8 = null,
         cache_account_history: ?[]const u8 = null,
         cache_grid: ?[]const u8 = null,
         account_count: usize = 10_000,
@@ -195,7 +195,7 @@ pub const Command = union(enum) {
         addresses_zero: bool,
         cache_accounts: u32,
         cache_transfers: u32,
-        cache_transfers_posted: u32,
+        cache_transfers_pending: u32,
         cache_account_history: u32,
         storage_size_limit: u64,
         cache_grid_blocks: u32,
@@ -218,7 +218,7 @@ pub const Command = union(enum) {
 
         cache_accounts: ?[]const u8 = null,
         cache_transfers: ?[]const u8 = null,
-        cache_transfers_posted: ?[]const u8 = null,
+        cache_transfers_pending: ?[]const u8 = null,
         cache_account_history: ?[]const u8 = null,
         cache_grid: ?[]const u8 = null,
         account_count: usize = 10_000,
@@ -331,7 +331,7 @@ pub fn parse_args(allocator: std.mem.Allocator, args_iterator: *std.process.ArgI
             const groove_config = StateMachine.Forest.groove_config;
             const AccountsValuesCache = groove_config.accounts.ObjectsCache.Cache;
             const TransfersValuesCache = groove_config.transfers.ObjectsCache.Cache;
-            const PostedValuesCache = groove_config.posted.ObjectsCache.Cache;
+            const TransfersPendingValuesCache = groove_config.transfers_pending.ObjectsCache.Cache;
             const AccountHistoryValuesCache = groove_config.account_history.ObjectsCache.Cache;
 
             const addresses = parse_addresses(allocator, start.addresses);
@@ -399,10 +399,10 @@ pub fn parse_args(allocator: std.mem.Allocator, args_iterator: *std.process.ArgI
                         TransfersValuesCache,
                         start.cache_transfers,
                     ),
-                    .cache_transfers_posted = parse_cache_size_to_count(
-                        StateMachine.PostedGrooveValue,
-                        PostedValuesCache,
-                        start.cache_transfers_posted,
+                    .cache_transfers_pending = parse_cache_size_to_count(
+                        StateMachine.TransferPending,
+                        TransfersPendingValuesCache,
+                        start.cache_transfers_pending,
                     ),
                     .cache_account_history = parse_cache_size_to_count(
                         StateMachine.AccountHistoryGrooveValue,
@@ -441,7 +441,7 @@ pub fn parse_args(allocator: std.mem.Allocator, args_iterator: *std.process.ArgI
                 .benchmark = .{
                     .cache_accounts = benchmark.cache_accounts,
                     .cache_transfers = benchmark.cache_transfers,
-                    .cache_transfers_posted = benchmark.cache_transfers_posted,
+                    .cache_transfers_pending = benchmark.cache_transfers_pending,
                     .cache_account_history = benchmark.cache_account_history,
                     .cache_grid = benchmark.cache_grid,
                     .account_count = benchmark.account_count,