--- a/src/tigerbeetle/main.zig
+++ b/src/tigerbeetle/main.zig
@@ -195,7 +195,7 @@ const Command = struct {
                 .lsm_forest_node_count = args.lsm_forest_node_count,
                 .cache_entries_accounts = args.cache_accounts,
                 .cache_entries_transfers = args.cache_transfers,
-                .cache_entries_posted = args.cache_transfers_posted,
+                .cache_entries_posted = args.cache_transfers_pending,
                 .cache_entries_account_history = args.cache_account_history,
             },
             .message_bus_options = .{