Unify PostedGroove and the index pending_status

As a last-minute change that affects storage stability, this PR moves the index `pending_status` from `TransfersGroove` to `PostedGroove` (now renamed to `TransfersPendingGroove`) changing the behavior of those objects:

- A `TransferPending` object is always inserted when a `Transfer` with `flags.pending == true` is created.
- Posting/voiding/expiring a pending transfer updates the entire `TransfersPending` object (no more manual updates to the `pending_status` index).

This change simplifies the mental model around tracking the outcome of a pending transfer and especially how we index its current status (while keeping the same searching capabilities).

As a downside, we can observe a _minor, but consistent_ performance penalty running the benchmark workload with the `--transfer-pending` flag (which marks 30% of all transfers as `pending` with a `timeout`, to be expired automatically).

Before
```
1244 batches in 63.13 s
load offered = 1000000 tx/s
load accepted = 158397 tx/s
batch latency p1 = 1 ms
batch latency p10 = 7 ms
batch latency p20 = 9 ms
batch latency p30 = 16 ms
batch latency p40 = 25 ms
batch latency p50 = 31 ms
batch latency p60 = 37 ms
batch latency p70 = 44 ms
batch latency p80 = 56 ms
batch latency p90 = 81 ms
batch latency p95 = 274 ms
batch latency p99 = 382 ms
batch latency p100 = 449 ms
```

After
```
1241 batches in 65.79 s
load offered = 1000000 tx/s
load accepted = 152004 tx/s
batch latency p1 = 2 ms
batch latency p10 = 8 ms
batch latency p20 = 10 ms
batch latency p30 = 18 ms
batch latency p40 = 26 ms
batch latency p50 = 32 ms
batch latency p60 = 40 ms
batch latency p70 = 49 ms
batch latency p80 = 63 ms
batch latency p90 = 81 ms
batch latency p95 = 284 ms
batch latency p99 = 373 ms
batch latency p100 = 446 ms
```